refactor(test_helper) move WebpackerHelpers to module
